+++ src/Tests/Aardvark.Base.Tests/Extensions/StreamTests.cs
@@ -3,7 +3,7 @@ using System;
 using System.Reflection;
 using System.IO;
 using Aardvark.Base;
-using System.Linq;
+using System.Runtime.ExceptionServices;
 
 namespace Aardvark.Tests.Extensions
 {
@@ -59,19 +59,9 @@ namespace Aardvark.Tests.Extensions
 
     static class StreamTests
     {
-        [Test]
-        public static void ReadBytes([Values(false, true)] bool netStandard20)
+        private static MethodInfo NetStandardReadBytesMethod
         {
-            var rnd = new RandomSystem();
-            var data = new byte[1234];
-            for (int i = 0; i < data.Length; i++) data[i] = (byte)rnd.UniformInt();
-
-            using TestStream s = new TestStream(data);
-            var result = new byte[57];
-            var offset = 3;
-            var count = result.Length - offset;
-
-            if (netStandard20)
+            get
             {
                 var outputDirectory = Path.GetDirectoryName(Assembly.GetExecutingAssembly().Location);
                 var assemblyPath = Path.GetFullPath(Path.Combine(outputDirectory!, "..", "netstandard2.0", "Aardvark.Base.dll"));
@@ -86,17 +76,87 @@ namespace Aardvark.Tests.Extensions
                 var asm = Assembly.LoadFile(assemblyPath);
                 var ext = asm.GetType($"Aardvark.Base.{nameof(StreamExtensions)}");
                 Type[] parameters = [typeof(Stream), typeof(byte[]), typeof(int), typeof(int)];
-                var mi = ext.GetMethod(nameof(StreamExtensions.ReadBytes), parameters);
+                return ext.GetMethod(nameof(StreamExtensions.ReadBytes), parameters);
+            }
+        }
 
-                mi.Invoke(null, [s, result, offset, count]);
+        private static void ReadBytes(bool netStandard20, Stream stream, byte[] buffer, int offset = 0, int count = -1)
+        {
+            if (netStandard20)
+            {
+                try
+                {
+                    NetStandardReadBytesMethod.Invoke(null, [stream, buffer, offset, count]);
+                }
+                catch (TargetInvocationException e) when (e.InnerException != null)
+                {
+                    ExceptionDispatchInfo.Capture(e.InnerException).Throw();
+                }
             }
             else
             {
-                s.ReadBytes(result, offset, count);
+                stream.ReadBytes(buffer, offset, count);
             }
+        }
+
+        [Test]
+        public static void ReadBytes([Values(false, true)] bool netStandard20)
+        {
+            var rnd = new RandomSystem();
+            var data = new byte[1234];
+            for (int i = 0; i < data.Length; i++) data[i] = (byte)rnd.UniformInt();
+
+            using TestStream s = new TestStream(data);
+            var result = new byte[57];
+            var offset = 3;
+            var count = result.Length - offset;
+
+            ReadBytes(netStandard20, s, result, offset, count);
 
             for (int i = 0; i < count; i++)
                 Assert.AreEqual(data[i], result[i + offset]);
+        }
+
+        [Test]
+        public static void ReadBytesRejectsNullStream([Values(false, true)] bool netStandard20)
+        {
+            var e = Assert.Throws<ArgumentNullException>(() => ReadBytes(netStandard20, null, new byte[1]));
+            Assert.AreEqual("stream", e.ParamName);
+        }
+
+        [Test]
+        public static void ReadBytesRejectsNullBufferWithDefaultCount([Values(false, true)] bool netStandard20)
+        {
+            using TestStream s = new TestStream([1]);
+
+            var e = Assert.Throws<ArgumentNullException>(() => ReadBytes(netStandard20, s, null));
+            Assert.AreEqual("buffer", e.ParamName);
+        }
+
+        [Test]
+        public static void ReadBytesRejectsNegativeOffset([Values(false, true)] bool netStandard20)
+        {
+            using TestStream s = new TestStream([1]);
+
+            var e = Assert.Throws<ArgumentOutOfRangeException>(() => ReadBytes(netStandard20, s, new byte[1], -1, 1));
+            Assert.AreEqual("offset", e.ParamName);
+        }
+
+        [Test]
+        public static void ReadBytesRejectsCountPastBufferEnd([Values(false, true)] bool netStandard20)
+        {
+            using TestStream s = new TestStream([1, 2]);
+
+            var e = Assert.Throws<ArgumentException>(() => ReadBytes(netStandard20, s, new byte[2], 1, 2));
+            Assert.AreEqual("count", e.ParamName);
+        }
+
+        [Test]
+        public static void ReadBytesThrowsEndOfStreamWhenSourceIsShort([Values(false, true)] bool netStandard20)
+        {
+            using TestStream s = new TestStream([1]);
+
+            Assert.Throws<EndOfStreamException>(() => ReadBytes(netStandard20, s, new byte[2], 0, 2));
         }
     }
 }